Move CLI tool functionalities to separate files.
This is a continuation of refactoring mender-artifact. Now all the
functionalities should be moved to separate files.

Changelog: None
