+++ cli/mender-artifact/artifacts.go
@@ -123,137 +123,6 @@ func (ae *artifactError) Error() string {
 	return ae.err.Error()
 }
 
-func checkIfValid(artifactPath string, key []byte) *artifactError {
-	verifyCallback := func(message, sig []byte) error {
-		return errors.New("artifact is signed but no verification key was provided")
-	}
-
-	if key != nil {
-		s := artifact.NewVerifier(key)
-		verifyCallback = s.Verify
-	}
-
-	// do not return error immediately if we can not validate signature;
-	// just continue checking consistency and return info if
-	// signature verification failed
-	var validationError error
-	ver := func(message, sig []byte) error {
-		if verifyCallback != nil {
-			if err := verifyCallback(message, sig); err != nil {
-				validationError = err
-			}
-		}
-		return nil
-	}
-
-	f, err := os.Open(artifactPath)
-	if err != nil {
-		return &artifactError{err: err}
-	}
-	defer f.Close()
-
-	ar := areader.NewReader(f)
-	_, err = read(ar, ver, nil)
-	if err != nil {
-		return &artifactError{err: err}
-	}
-
-	if validationError != nil {
-		return &artifactError{
-			err: fmt.Errorf("artifact file '%s' formatted correctly, "+
-				"but error validating signature: %s", artifactPath, validationError),
-			badSignature: true,
-		}
-	}
-	return nil
-}
-
-func validateArtifact(c *cli.Context) error {
-	if c.NArg() == 0 {
-		return cli.NewExitError("Nothing specified, nothing validated. \nMaybe you wanted"+
-			" to say 'artifacts validate <pathspec>'?", 1)
-	}
-
-	var key []byte
-	var err error
-	if c.String("key") != "" {
-		key, err = getKey(c.String("key"))
-		if err != nil {
-			return cli.NewExitError("Can not read key: "+err.Error(), 1)
-		}
-	}
-
-	if err := checkIfValid(c.Args().First(), key); err != nil {
-		return cli.NewExitError(err.Error(), 1)
-	}
-
-	fmt.Printf("Artifact file '%s' validated successfully\n", c.Args().First())
-	return nil
-}
-
-func signExisting(c *cli.Context) error {
-	if c.NArg() == 0 {
-		return cli.NewExitError("Nothing specified, nothing signed. \nMaybe you wanted"+
-			" to say 'artifacts sign <pathspec>'?", 1)
-	}
-
-	if len(c.String("key")) == 0 {
-		return cli.NewExitError("Missing signing key; "+
-			"please use `-k` parameter for providing one", 1)
-	}
-
-	privateKey, err := getKey(c.String("key"))
-	if err != nil {
-		return cli.NewExitError("Can not use signing key provided: "+err.Error(), 1)
-	}
-
-	tFile, err := ioutil.TempFile("", "mender-artifact")
-	if err != nil {
-		return errors.Wrap(err,
-			"Can not create temporary file for storing artifact")
-	}
-	defer os.Remove(tFile.Name())
-	defer tFile.Close()
-
-	f, err := os.Open(c.Args().First())
-	if err != nil {
-		return errors.Wrapf(err, "Can not open: %s", c.Args().First())
-	}
-	defer f.Close()
-
-	reader, err := repack(f, tFile, privateKey, "", "")
-	if err != nil {
-		return err
-	}
-
-	switch ver := reader.GetInfo().Version; ver {
-	case 1:
-		return cli.NewExitError("Can not sign v1 artifact", 1)
-	case 2:
-		if reader.IsSigned && !c.Bool("force") {
-			return cli.NewExitError("Trying to sign already signed artifact; "+
-				"please use force option", 1)
-		}
-	default:
-		return cli.NewExitError("Unsupported version of artifact file: "+string(ver), 1)
-	}
-
-	if err = tFile.Close(); err != nil {
-		return err
-	}
-
-	name := c.Args().First()
-	if len(c.String("output-path")) > 0 {
-		name = c.String("output-path")
-	}
-
-	err = os.Rename(tFile.Name(), name)
-	if err != nil {
-		return cli.NewExitError("Can not store signed artifact: "+err.Error(), 1)
-	}
-	return nil
-}
-
 func unpackArtifact(name string) (string, error) {
 	f, err := os.Open(name)
 	if err != nil {
@@ -630,58 +499,4 @@ func getCandidatesForModify(path string, key []byte) ([]partition, bool, error) 
 		modifyCandidates = append(modifyCandidates, parts...)
 	}
 	return modifyCandidates, isArtifact, nil
-}
-
-func modifyArtifact(c *cli.Context) error {
-	if c.NArg() == 0 {
-		return cli.NewExitError("Nothing specified, nothing will be modified. \n"+
-			"Maybe you wanted to say 'artifacts read <pathspec>'?", 1)
-	}
-
-	if _, err := os.Stat(c.Args().First()); err != nil && os.IsNotExist(err) {
-		return cli.NewExitError("File ["+c.Args().First()+"] does not exist.", 1)
-	}
-
-	pubKey, err := processModifyKey(c.String("key"))
-	if err != nil {
-		return cli.NewExitError("Error processing private key: "+err.Error(), 1)
-	}
-
-	modifyCandidates, isArtifact, err :=
-		getCandidatesForModify(c.Args().First(), pubKey)
-
-	if err != nil {
-		return cli.NewExitError("Error selecting images for modification: "+err.Error(), 1)
-	}
-
-	if len(modifyCandidates) > 1 || isArtifact {
-		for _, mc := range modifyCandidates {
-			defer os.Remove(mc.path)
-		}
-	}
-
-	for _, toModify := range modifyCandidates {
-		if err := modifyExisting(c, toModify.path); err != nil {
-			return cli.NewExitError("Error modifying artifact["+toModify.path+"]: "+
-				err.Error(), 1)
-		}
-	}
-
-	if len(modifyCandidates) > 1 {
-		// make modified images part of sdimg again
-		if err := repackSdimg(modifyCandidates, c.Args().First()); err != nil {
-			return cli.NewExitError("Can not recreate sdimg file: "+err.Error(), 1)
-		}
-		return nil
-	}
-
-	if isArtifact {
-		// re-create the artifact
-		err := repackArtifact(c.Args().First(), modifyCandidates[0].path,
-			c.String("key"), c.String("name"))
-		if err != nil {
-			return cli.NewExitError("Can not recreate artifact: "+err.Error(), 1)
-		}
-	}
-	return nil
 }
