+++ cli/mender-artifact/artifacts_test.go
@@ -89,58 +89,6 @@ func WriteArtifact(dir string, ver int, update string) error {
 		"mender-1.1", updates, nil)
 }
 
-func TestArtifactsWrite(t *testing.T) {
-	os.Args = []string{"mender-artifact", "write"}
-	err := run()
-	// should output help message and no error
-	assert.NoError(t, err)
-
-	os.Args = []string{"mender-artifact", "write", "rootfs-image"}
-	err = run()
-	assert.Error(t, err)
-	assert.Equal(t, 1, lastExitCode)
-	assert.Equal(t, "must provide `device-type`, `artifact-name` and `update`\n",
-		fakeErrWriter.String())
-
-	updateTestDir, _ := ioutil.TempDir("", "update")
-	defer os.RemoveAll(updateTestDir)
-
-	err = MakeFakeUpdateDir(updateTestDir,
-		[]TestDirEntry{
-			{
-				Path:    "update.ext4",
-				Content: []byte("my update"),
-				IsDir:   false,
-			},
-		})
-	assert.NoError(t, err)
-
-	// no whitespace allowed in artifact-name
-	os.Args = []string{"mender-artifact", "write", "rootfs-image", "-t", "my-device",
-		"-n", "mender-1. 1", "-u", filepath.Join(updateTestDir, "update.ext4"),
-		"-o", filepath.Join(updateTestDir, "art.mender")}
-	err = run()
-	assert.Equal(t, "whitespace is not allowed in the artifact-name", err.Error())
-
-	// store named file
-	os.Args = []string{"mender-artifact", "write", "rootfs-image", "-t", "my-device",
-		"-n", "mender-1.1", "-u", filepath.Join(updateTestDir, "update.ext4"),
-		"-o", filepath.Join(updateTestDir, "art.mender")}
-	err = run()
-	assert.NoError(t, err)
-
-	fs, err := os.Stat(filepath.Join(updateTestDir, "art.mender"))
-	assert.NoError(t, err)
-	assert.False(t, fs.IsDir())
-
-	// store named file
-	os.Args = []string{"mender-artifact", "write", "rootfs-image", "-t", "my-device",
-		"-n", "mender-1.1", "-u", filepath.Join(updateTestDir, "update.ext4"),
-		"-o", filepath.Join(updateTestDir, "art.mender"), "-v", "3"}
-	err = run()
-	assert.Error(t, err)
-}
-
 func generateKeys() ([]byte, []byte, error) {
 	// key size needs to be 512 bits to handle message length
 	priv, err := rsa.GenerateKey(rand.Reader, 512)
@@ -267,286 +215,6 @@ func TestArtifactsSigned(t *testing.T) {
 	err = run()
 	assert.Error(t, err)
 	assert.Equal(t, "can not use signed artifact with version 1\n",
-		fakeErrWriter.String())
-}
-
-func TestSignExistingV1(t *testing.T) {
-	// first create archive, that we will be able to read
-	updateTestDir, _ := ioutil.TempDir("", "update")
-	defer os.RemoveAll(updateTestDir)
-
-	priv, pub, err := generateKeys()
-	assert.NoError(t, err)
-
-	err = WriteArtifact(updateTestDir, 1, "")
-	assert.NoError(t, err)
-
-	err = MakeFakeUpdateDir(updateTestDir,
-		[]TestDirEntry{
-			{
-				Path:    "private.key",
-				Content: priv,
-				IsDir:   false,
-			},
-			{
-				Path:    "public.key",
-				Content: pub,
-				IsDir:   false,
-			},
-		})
-	assert.NoError(t, err)
-
-	os.Args = []string{"mender-artifact", "sign",
-		"-k", filepath.Join(updateTestDir, "private.key"),
-		filepath.Join(updateTestDir, "artifact.mender")}
-
-	err = run()
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "can not create version 1 signed artifact")
-}
-
-func TestSignExistingV2(t *testing.T) {
-	// first create archive, that we will be able to read
-	updateTestDir, _ := ioutil.TempDir("", "update")
-	defer os.RemoveAll(updateTestDir)
-
-	priv, pub, err := generateKeys()
-	assert.NoError(t, err)
-
-	err = WriteArtifact(updateTestDir, 2, "")
-	assert.NoError(t, err)
-
-	err = MakeFakeUpdateDir(updateTestDir,
-		[]TestDirEntry{
-			{
-				Path:    "private.key",
-				Content: priv,
-				IsDir:   false,
-			},
-			{
-				Path:    "public.key",
-				Content: pub,
-				IsDir:   false,
-			},
-		})
-	assert.NoError(t, err)
-
-	os.Args = []string{"mender-artifact", "sign",
-		"-k", filepath.Join(updateTestDir, "private.key"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender.sig"),
-		filepath.Join(updateTestDir, "artifact.mender")}
-
-	err = run()
-	assert.NoError(t, err)
-
-	os.Args = []string{"mender-artifact", "validate",
-		"-k", filepath.Join(updateTestDir, "public.key"),
-		filepath.Join(updateTestDir, "artifact.mender.sig")}
-
-	err = run()
-	assert.NoError(t, err)
-
-	// now check if signing already signed will fail
-	os.Args = []string{"mender-artifact", "sign",
-		"-k", filepath.Join(updateTestDir, "private.key"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender.sig"),
-		filepath.Join(updateTestDir, "artifact.mender.sig")}
-	err = run()
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "Trying to sign already signed artifact")
-
-	// and the same as above with force option
-	os.Args = []string{"mender-artifact", "sign",
-		"-k", filepath.Join(updateTestDir, "private.key"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender.sig"), "-f",
-		filepath.Join(updateTestDir, "artifact.mender.sig")}
-	err = run()
-	assert.NoError(t, err)
-}
-
-func TestSignExistingWithScripts(t *testing.T) {
-	updateTestDir, _ := ioutil.TempDir("", "update")
-	defer os.RemoveAll(updateTestDir)
-
-	priv, pub, err := generateKeys()
-	assert.NoError(t, err)
-
-	err = MakeFakeUpdateDir(updateTestDir,
-		[]TestDirEntry{
-			{
-				Path:    "private.key",
-				Content: priv,
-				IsDir:   false,
-			},
-			{
-				Path:    "public.key",
-				Content: pub,
-				IsDir:   false,
-			},
-			{
-				Path:    "update.ext4",
-				Content: []byte("my update"),
-				IsDir:   false,
-			},
-			{
-				Path:    "ArtifactInstall_Enter_99",
-				Content: []byte("this is first enter script"),
-				IsDir:   false,
-			},
-			{
-				Path:    "ArtifactInstall_Leave_01",
-				Content: []byte("this is leave script"),
-				IsDir:   false,
-			},
-		})
-	assert.NoError(t, err)
-
-	// write artifact
-	os.Args = []string{"mender-artifact", "write", "rootfs-image", "-t", "my-device",
-		"-n", "mender-1.1", "-u", filepath.Join(updateTestDir, "update.ext4"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender"),
-		"-s", filepath.Join(updateTestDir, "ArtifactInstall_Enter_99"),
-		"-s", filepath.Join(updateTestDir, "ArtifactInstall_Leave_01")}
-	err = run()
-	assert.NoError(t, err)
-
-	// test sign exisiting
-	os.Args = []string{"mender-artifact", "sign",
-		"-k", "-o", filepath.Join(updateTestDir, "artifact.mender.sig"),
-		filepath.Join(updateTestDir, "artifact.mender")}
-
-	err = run()
-	assert.Error(t, err)
-
-	// test sign exisiting
-	os.Args = []string{"mender-artifact", "sign",
-		"-o", filepath.Join(updateTestDir, "artifact.mender.sig"),
-		filepath.Join(updateTestDir, "artifact.mender")}
-
-	err = run()
-	assert.Error(t, err)
-
-	// test sign exisiting
-	os.Args = []string{"mender-artifact", "sign",
-		"-k", filepath.Join(updateTestDir, "private.key"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender.sig"),
-		filepath.Join(updateTestDir, "artifact.mender")}
-
-	err = run()
-	assert.NoError(t, err)
-
-}
-
-func TestArtifactsValidateError(t *testing.T) {
-	os.Args = []string{"mender-artifact", "validate"}
-	err := run()
-	assert.Error(t, err)
-	assert.Contains(t, errors.Cause(err).Error(),
-		"Nothing specified, nothing validated.")
-
-	os.Args = []string{"mender-artifact", "validate", "non-existing"}
-	fakeErrWriter.Reset()
-	err = run()
-	assert.Error(t, err)
-	assert.Equal(t, 1, lastExitCode)
-	assert.Contains(t, fakeErrWriter.String(), "no such file")
-}
-
-func TestArtifactsValidate(t *testing.T) {
-	// first create archive, that we will be able to read
-	updateTestDir, _ := ioutil.TempDir("", "update")
-	defer os.RemoveAll(updateTestDir)
-
-	err := WriteArtifact(updateTestDir, 1, "")
-	assert.NoError(t, err)
-
-	os.Args = []string{"mender-artifact", "validate",
-		filepath.Join(updateTestDir, "artifact.mender")}
-	err = run()
-	assert.NoError(t, err)
-}
-
-func TestWithScripts(t *testing.T) {
-	updateTestDir, _ := ioutil.TempDir("", "update")
-	defer os.RemoveAll(updateTestDir)
-
-	err := MakeFakeUpdateDir(updateTestDir,
-		[]TestDirEntry{
-			{
-				Path:    "update.ext4",
-				Content: []byte("my update"),
-				IsDir:   false,
-			},
-			{
-				Path:    "ArtifactInstall_Enter_99",
-				Content: []byte("this is first enter script"),
-				IsDir:   false,
-			},
-			{
-				Path:    "ArtifactInstall_Leave_01",
-				Content: []byte("this is leave script"),
-				IsDir:   false,
-			},
-			{
-				Path:    "script-dir",
-				Content: []byte(""),
-				IsDir:   true,
-			},
-			{
-				Path:    "script-dir/ArtifactReboot_Enter_99",
-				Content: []byte("this is reboot enter script"),
-				IsDir:   false,
-			},
-			{
-				Path:    "script-dir/ArtifactReboot_Leave_01",
-				Content: []byte("this is reboot leave script"),
-				IsDir:   false,
-			},
-			{
-				Path:    "InvalidScript",
-				Content: []byte("this is invalid script"),
-				IsDir:   false,
-			},
-		})
-	assert.NoError(t, err)
-
-	// write artifact
-	os.Args = []string{"mender-artifact", "write", "rootfs-image", "-t", "my-device",
-		"-n", "mender-1.1", "-u", filepath.Join(updateTestDir, "update.ext4"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender"),
-		"-s", filepath.Join(updateTestDir, "ArtifactInstall_Enter_99"),
-		"-s", filepath.Join(updateTestDir, "ArtifactInstall_Leave_01"),
-		"-s", filepath.Join(updateTestDir, "script-dir")}
-	err = run()
-	assert.NoError(t, err)
-
-	// read artifact
-	os.Args = []string{"mender-artifact", "read",
-		filepath.Join(updateTestDir, "artifact.mender")}
-	err = run()
-	assert.NoError(t, err)
-
-	// write artifact vith invalid version
-	os.Args = []string{"mender-artifact", "write", "rootfs-image", "-t", "my-device",
-		"-n", "mender-1.1", "-u", filepath.Join(updateTestDir, "update.ext4"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender"),
-		"-s", filepath.Join(updateTestDir, "ArtifactInstall_Enter_99"),
-		"-v", "1"}
-	fakeErrWriter.Reset()
-	err = run()
-	assert.Error(t, err)
-	assert.Equal(t, "can not use scripts artifact with version 1\n",
-		fakeErrWriter.String())
-
-	// write artifact vith invalid script name
-	os.Args = []string{"mender-artifact", "write", "rootfs-image", "-t", "my-device",
-		"-n", "mender-1.1", "-u", filepath.Join(updateTestDir, "update.ext4"),
-		"-o", filepath.Join(updateTestDir, "artifact.mender"),
-		"-s", filepath.Join(updateTestDir, "InvalidScript")}
-	fakeErrWriter.Reset()
-	err = run()
-	assert.Error(t, err)
-	assert.Equal(t, "scripter: invalid script: InvalidScript\n",
 		fakeErrWriter.String())
 }
