+++ cli/mender-artifact/validate_test.go
@@ -1,0 +1,54 @@
+// Copyright 2018 Northern.tech AS
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//        http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+
+package main
+
+import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/pkg/errors"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestArtifactsValidate(t *testing.T) {
+	// first create archive, that we will be able to read
+	updateTestDir, _ := ioutil.TempDir("", "update")
+	defer os.RemoveAll(updateTestDir)
+
+	err := WriteArtifact(updateTestDir, 1, "")
+	assert.NoError(t, err)
+
+	os.Args = []string{"mender-artifact", "validate",
+		filepath.Join(updateTestDir, "artifact.mender")}
+	err = run()
+	assert.NoError(t, err)
+}
+
+func TestArtifactsValidateError(t *testing.T) {
+	os.Args = []string{"mender-artifact", "validate"}
+	err := run()
+	assert.Error(t, err)
+	assert.Contains(t, errors.Cause(err).Error(),
+		"Nothing specified, nothing validated.")
+
+	os.Args = []string{"mender-artifact", "validate", "non-existing"}
+	fakeErrWriter.Reset()
+	err = run()
+	assert.Error(t, err)
+	assert.Equal(t, 1, lastExitCode)
+	assert.Contains(t, fakeErrWriter.String(), "no such file")
+}
